+++ CHANGELOG.md
@@ -24,6 +24,13 @@ Changes are grouped as follows
 - Add a `load` method to CogniteClient, ClientConfig, and CredenitalProvider (and all it's subclasses).
 - Add `apply_settings` method to `global_config` to pass in a dict of settings
 
+## [7.54.12] - 2024-08-08
+### Fixed
+- NodeList and EdgeList (and subclasses) now expects an instance ID, `(space, external_id)` in the `.get` method.
+  Using just an `external_id` is still possible, but deprecated as it is ambiguous in the absence of the space
+  identifier, and will just return the last matching instance (as previously).
+- SpaceList.get now works and expects a space identifier in the `.get` method.
+
 ## [7.54.11] - 2024-07-26
 ### Fixed
 - Creating a Group with an `UnknownAcl` supported by the API no longer raises a client-side `ValueError` after